LayoutTests/platform/mac/accessibility/aria-valuetext.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="slider" role="slider" aria-valuetext="MAX" aria-valuenow="100" aria-valuemin="0" aria-valuemax="100">slider</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that aria-valuetext reachable as the AXValueDescription");

    if (window.accessibilityController) {

          // this text field should be required.
          document.getElementById("body").focus();
          var obj = accessibilityController.focusedElement.childAtIndex(0);
          var succeeded = obj.valueDescription == "MAX";
          shouldBe("succeeded", "true");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
