LayoutTests/platform/mac/accessibility/aria-valuetext-expected.txt

slider
This tests that aria-valuetext reachable as the AXValueDescription

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
